+++ src/main/java/com/example/demo/service/impl/ManagementService.java
@@ -1,10 +1,14 @@
 package com.example.demo.service.impl;
 
+import com.example.demo.common.InformationException;
 import com.example.demo.common.Result;
 import com.example.demo.dao.IInforDao;
+import com.example.demo.dao.IUserDao;
 import com.example.demo.dao.IVolunteeringDao;
+import com.example.demo.pojo.User;
 import com.example.demo.pojo.Volunteering;
 import com.example.demo.service.IManagementService;
+import com.example.demo.service.IUserService;
 import com.github.pagehelper.PageHelper;
 import lombok.Synchronized;
 import lombok.extern.java.Log;
@@ -17,6 +21,7 @@ import org.springframework.transaction.TransactionDefinition;
 import org.springframework.transaction.TransactionStatus;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.sound.sampled.Line;
 import java.sql.SQLException;
 import java.util.List;
 
@@ -27,6 +32,8 @@ public class ManagementService implements IManagementService {
     private IInforDao iInforDao;
     @Autowired
     private IVolunteeringDao iVolunteeringDao;
+    @Autowired
+    private IUserDao iUserDao;
     // 事务管理
     @Autowired
     private PlatformTransactionManager platformTransactionManager;
@@ -49,30 +56,12 @@ public class ManagementService implements IManagementService {
 
     @Override
     public Result<?> getInforByUid(Integer uid) {
-        Result<?> result;
-        try{
-            result = Result.success(iInforDao.getInforByUid(uid));
-            return result;
-        } catch (DataAccessException e){
-            result = Result.error(Result.INFORMATION_ERROR,"传入参数未知");
-        } catch (Exception e){
-            result = Result.error(Result.UNKNOWN_ERROR, e.getMessage());
-        }
-        return result;
+        return Result.success(iInforDao.getInforByUid(uid));
     }
 
     @Override
     public Result<?> getInforByVid(Integer vid) {
-        Result<?> result;
-        try{
-            result = Result.success(iInforDao.getInforByVid(vid));
-            return result;
-        } catch (DataAccessException e){
-            result = Result.error(Result.INFORMATION_ERROR,"传入参数未知");
-        } catch (Exception e){
-            result = Result.error(Result.UNKNOWN_ERROR, e.getMessage());
-        }
-        return result;
+        return Result.success(iInforDao.getInforByVid(vid));
     }
 
     @Transactional
@@ -80,11 +69,16 @@ public class ManagementService implements IManagementService {
     public Result<?> deleteSpecifiedInfor(Integer uid, Integer vid) {
         Result<?> result;
         try{
+            if (iUserDao.getUserById(uid).size() == 0){
+                throw new InformationException("没有该用户");
+            }
+            if(iVolunteeringDao.getVolById(vid).size() == 0){
+                throw new InformationException("没有该活动");
+            }
             iInforDao.deleteSpecifiedInfor(uid,vid);
             result = Result.success();
-            return result;
-        } catch (DataAccessException e){
-            result = Result.error(Result.INFORMATION_ERROR,"信息错误");
+        } catch (InformationException e){
+            result = Result.error(Result.INFORMATION_ERROR,e.getMessage());
         } catch (Exception e){
             result = Result.error(Result.UNKNOWN_ERROR, e.getMessage());
         }
@@ -96,13 +90,15 @@ public class ManagementService implements IManagementService {
     public Result<?> deleteInforByVid(Integer vid) {
         Result<?> result;
         try{
+            if(iVolunteeringDao.getVolById(vid).size() == 0){
+                throw new InformationException("没有该活动");
+            }
             iInforDao.deleteInforByVid(vid);
             result = Result.success();
-            return result;
-        } catch (DataAccessException e){
-            result = Result.error(Result.INFORMATION_ERROR,"信息错误");
+        } catch (InformationException e){
+            result = Result.error(Result.INFORMATION_ERROR,e.getMessage());
         } catch (Exception e){
-            result = Result.error(Result.UNKNOWN_ERROR, e.getMessage());
+            result = Result.error(Result.UNKNOWN_ERROR, "未知错误");
         }
         return result;
     }
@@ -111,32 +107,33 @@ public class ManagementService implements IManagementService {
     @Override
     public synchronized Result<?> registration(Integer uid, Integer vid) {
         // 开启事务
-        log.info("uid="+uid+"我开始执行啦！");
         TransactionStatus transactionStatus = platformTransactionManager.getTransaction(transactionDefinition);
         Result<?> result;
         try{
+            if (iUserDao.getUserById(uid).size() == 0){
+                throw new InformationException("没有该用户");
+            }
+
             Integer currentNum = iInforDao.getCountByVid(vid);
             List<Volunteering> vList = iVolunteeringDao.getVolById(vid);
             if(vList.size() == 0){
-                throw new Exception("没有该活动");
+                throw new InformationException("没有该活动");
             }
             Volunteering v = vList.get(0);
             if(v.getPNum().equals(currentNum)){
-                result = Result.error(Result.FULL_REGISTRATION,"人数已满，无法报名");
-            } else {
-                iInforDao.addInfor(uid,vid);
-                result = Result.success();
+                throw new InformationException("人数已满，无法报名");
             }
+
+            iInforDao.addInfor(uid,vid);
+            result = Result.success();
             platformTransactionManager.commit(transactionStatus);
             return result;
-        } catch (DataAccessException e){
-            result = Result.error(Result.INFORMATION_ERROR,"信息错误");
+        } catch (InformationException e){
+            result = Result.error(Result.INFORMATION_ERROR,e.getMessage());
             platformTransactionManager.rollback(transactionStatus);
         } catch (Exception e){
             result = Result.error(Result.UNKNOWN_ERROR, e.getMessage());
             platformTransactionManager.rollback(transactionStatus);
-        } finally {
-            log.info("uid="+uid+"执行完毕了！！！！");
         }
         return result;
     }
